* fixed new brightness_contrast_gamma function to work on 16bit data 	* introduced new getRGB/setRGB with double arguments for more correct 	  and comfortable handling 	* needed to explicitly cast some other code places as the with the 	  overloaded getRGB/setRGB the code became ambegious

+++ lib/segmentation.cc
@@ -64,7 +64,7 @@ bool Segment::Subdivide(const FGMatrix& img, double tolerance, unsigned int min_
 }
 
 
-void Segment::Draw(Image& output, unsigned int r, unsigned int g, unsigned int b)
+void Segment::Draw(Image& output, uint16_t r, uint16_t g, uint16_t b)
 {
   Image::iterator color = output.begin();
   color.setRGB (r, g, b);
